test: 통합 테스트를 integration 패키지로 네이밍을 바꾸고 Nested 구조로 변경
